type_a/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-anthropic>=0.0.1
langchain-groq>=0.0.1
langgraph>=0.0.15
crewai>=0.11.0
neo4j>=5.14.1
pgvector>=0.2.3
docker-compose>=1.29.2
lightrag>=0.1.0
ollama>=0.1.6
nginx>=1.25.3
langfuse>=2.0.0
apache-tika>=2.9.1
PyMuPDF>=1.23.8
PyPDF2>=3.0.1
python-docx>=1.0.1
beautifulsoup4>=4.12.2
requests>=2.31.0
pandas>=2.1.4
numpy>=1.26.2
pydantic>=2.5.2
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.1
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1 